fix dir path
