+++ kkanbu/board/admin.py
