+++ kkanbu/operation/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class BoardConfig(AppConfig):
+class OperationConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "board"
+    name = "kkanbu.operation"
