+++ kkanbu/operation/tests.py
